requirements.txt
flask
numpy
faiss
flask_sqlalchemy
flask_restful
passlib
pyjwt
gunicorn